history_service/requirements.txt
Django==5.1
djangorestframework==3.15
django-celery-beat==2.8.1
psycopg2-binary==2.9.9
python-decouple==3.8
confluent-kafka==2.12.0
celery==5.4.0
redis==5.0.8
requests==2.32.5
# optional but recommended for observability
django-prometheus==2.3.1
opentelemetry-distro==0.46b0
# locust==2.40.0
